Non-matching decomp of func_80101EF4_110EA4

+++ include/functions.us.h
@@ -1076,7 +1076,7 @@ void func_800DDE54_ECE04(u8 arg0, s8 arg1);
 void func_800DDE90_ECE40(u8 arg0, s8 arg1, s8 arg2);
 void func_800DDD30_ECCE0(u8 arg0, s16 arg1, s16 arg2, s16 arg3);
 void func_800DDF18_ECEC8(u8 arg0);
-void func_80101EF4_110EA4(u8 arg0, s16 arg1, s16 arg2, s16 arg3, s32 arg4, s32 arg5, s32 arg6);
+void func_80101EF4_110EA4(s16 arg0, s16 arg1, s16 arg2, s16 arg3, s16 arg4, s16 arg5, s16 arg6);
 void func_80097CB4_A6C64(UnkA6C64Keyframe *arg0, UnkA6C64Keyframe *arg1, UnkA6C64Keyframe *arg2, f32 arg3);
 void func_8000C81C_D41C(s32 *arg0, s16 *arg1, s16 *arg2, s32 *arg3);
 void func_8000CC3C_D83C(AnimChannelState *arg0, s32 arg1);
